# post-data-management-back-end-development/docs/swagger.yaml
basePath: /v1
definitions:
  apierrors.APIErrorResponse:
    properties:
      error:
        $ref: '#/definitions/apierrors.AppError'
      message:
        description: Descriptive message providing details about the error.
        type: string
      status_code:
        description: HTTP status code representing the error type (e.g., 404 for Not
          Found).
        type: integer
      success:
        description: Success flag to indicate if the operation was successful.
        type: boolean
    type: object
  apierrors.AppError:
    properties:
      code:
        type: string
      field_errors:
        items:
          $ref: '#/definitions/apierrors.FieldError'
        type: array
      id:
        type: string
      message:
        type: string
    type: object
...
  handler.EstablishmentRegisterRequest:
    properties:
      created_by:
        type: string
      created_date:
        type: string
      establishment_register_name:
        type: string
      office_id:
        type: integer
      office_name:
        type: string
      status:
        type: string
    required:
    - created_by
    - created_date
    - establishment_register_name
    - office_id
    - office_name
    - status
    type: object
...
  handler.PostManagementMakerCreateRequests:
    properties:
      postcreatereq:
        items:
          $ref: '#/definitions/handler.PostManagementMakerCreateRequest'
        type: array
    type: object
...
  response.CreateEstablishmentRegisterAPIResponse:
    properties:
      data:
        $ref: '#/definitions/response.CreateEstablishmentRegisterResponse'
      message:
        type: string
      status_code:
        type: integer
      success:
        type: boolean
    type: object
...
  response.CreatePostManagementMakerAPIResponse:
    properties:
      data:
        items:
          $ref: '#/definitions/response.CreatePostManagementMakerResponse'
        type: array
      message:
        type: string
      status_code:
        type: integer
      success:
        type: boolean
    type: object
  response.CreatePostManagementMakerResponse:
    properties:
      office_id:
        type: integer
      office_name:
        type: string
      post_id:
        type: integer
      post_management_maker_id:
        type: integer
      post_name:
        type: string
      status:
        type: string
    type: object
...
  response.FetchEstablishmentRegisterAPIResponse:
    properties:
      data:
        items:
          $ref: '#/definitions/response.FetchEstablishmentRegisterResponse'
        type: array
      limit:
        type: integer
      message:
        type: string
      order_by:
        type: string
      returned_records_count:
        type: integer
      skip:
        type: integer
      sort_type:
        type: string
      status_code:
        type: integer
      success:
        type: boolean
      total_records_count:
        type: integer
    type: object
  response.FetchEstablishmentRegisterResponse:
    properties:
      cadre_id:
        type: integer
      cadre_name:
        type: string
      designation:
        type: string
      employee_group:
        type: string
      establishment_register_id:
        type: integer
      grade_pay:
        type: integer
      group_id:
        type: integer
      group_name:
        type: string
      office_id:
        type: integer
      office_name:
        type: string
      pay_level:
        type: integer
      permanent_status:
        type: boolean
      post_id:
        type: integer
      post_name:
        type: string
      remarks:
        type: string
      status:
        type: string
    type: object
...
  response.FetchPostsByOfficeIDAndMakerAPIResponse:
    properties:
      data:
        items:
          $ref: '#/definitions/response.FetchPostsByOfficeIDAndMakerResponse'
        type: array
      limit:
        type: integer
      message:
        type: string
      order_by:
        type: string
      returned_records_count:
        type: integer
      skip:
        type: integer
      sort_type:
        type: string
      status_code:
        type: integer
      success:
        type: boolean
      total_records_count:
        type: integer
    type: object
...
  response.ListAndFilterDesignationsAPIResponse:
    properties:
      data:
        items:
          $ref: '#/definitions/response.ListAndFilterDesignationsResponse'
        type: array
      limit:
        type: integer
      message:
        type: string
      order_by:
        type: string
      returned_records_count:
        type: integer
      skip:
        type: integer
      sort_type:
        type: string
      status_code:
        type: integer
      success:
        type: boolean
      total_records_count:
        type: integer
    type: object
  response.ListAndFilterDesignationsResponse:
    properties:
      cadre_id:
        type: integer
      cadre_name:
        type: string
      designation:
        type: string
      designation_id:
        type: integer
      designation_name:
        type: string
      group_id:
        type: integer
      group_name:
        type: string
    type: object
...
  response.ListCadresAPIResponse:
    properties:
      data:
        items:
          $ref: '#/definitions/response.ListCadresResponse'
        type: array
      limit:
        type: integer
      message:
        type: string
      order_by:
        type: string
      returned_records_count:
        type: integer
      skip:
        type: integer
      sort_type:
        type: string
      status_code:
        type: integer
      success:
        type: boolean
      total_records_count:
        type: integer
    type: object
  response.ListCadresResponse:
    properties:
      cadre_id:
        type: integer
      cadre_name:
        type: string
      group_name:
        type: string
    type: object
...
  response.PostManagementByOfficeAndPostAPIResponse:
    properties:
      data:
        items:
          $ref: '#/definitions/response.PostManagementByOfficeAndPostResponse'
        type: array
      limit:
        type: integer
      message:
        type: string
      order_by:
        type: string
      returned_records_count:
        type: integer
      skip:
        type: integer
      sort_type:
        type: string
      status_code:
        type: integer
      success:
        type: boolean
      total_records_count:
        type: integer
    type: object
...
  response.PostManagementByOfficeIDResponse1:
    properties:
      data:
        items:
          $ref: '#/definitions/response.VacantPostResponse'
        type: array
      limit:
        type: integer
      message:
        type: string
      order_by:
        type: string
      returned_records_count:
        type: integer
      skip:
        type: integer
      sort_type:
        type: string
      status_code:
        type: integer
      success:
        type: boolean
      total_records_count:
        type: integer
    type: object
...
  response.VacantPostResponse:
    properties:
      post_id:
        type: integer
    type: object
host: localhost:8080
info:
  contact:
    email: support_cept@indiapost.gov.in
    name: API Support Team
    url: http://cept.gov.in/support
  description: A comprehensive API for managing Post Management Data Management, offering
    endpoints for creation, update, deletion, and retrieval of Posts Data.
  license:
    name: Apache 2.0
    url: http://www.apache.org/licenses/LICENSE-2.0.html
  termsOfService: http://cept.gov.in/terms
  title: Post Management Data Management API
  version: "1.0"
paths:
...
  /office-post-details:
    get:
      consumes:
      - application/json
      description: Fetch all vacant posts in a given office. If no posts or no vacant
        posts are found, a message will be returned.
      parameters:
      - description: Office ID
        in: query
        name: office-id
        required: true
        type: integer
      - description: Number of records to skip (for pagination)
        in: query
        name: skip
        type: integer
      - description: Number of records to return (for pagination)
        in: query
        name: limit
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: Posts fetched successfully
          schema:
            $ref: '#/definitions/response.PostManagementByOfficeIDResponse1'
        "400":
          description: Validation or binding error
          schema:
            $ref: '#/definitions/apierrors.APIErrorResponse'
        "500":
          description: Internal server error
          schema:
            $ref: '#/definitions/apierrors.APIErrorResponse'
      summary: Get Post Management details by Office ID
      tags:
      - Post Management
  /post-management/cadres:
    get:
      consumes:
      - application/json
      description: List all cadre types or filter by group-code if provided
      parameters:
      - description: Filter cadres by group code
        in: query
        name: group-code
        type: string
      - in: query
        name: limit
        type: integer
      - in: query
        name: order_by
        type: string
      - in: query
        name: skip
        type: integer
      - in: query
        name: sort_type
        type: string
      - in: query
        name: total_records_required
        type: boolean
      produces:
      - application/json
      responses:
        "200":
          description: Cadre list retrieved successfully
          schema:
            $ref: '#/definitions/response.ListCadresAPIResponse'
        "400":
          description: Validation error
          schema:
            $ref: '#/definitions/apierrors.APIErrorResponse'
        "401":
          description: Unauthorized error
          schema:
            $ref: '#/definitions/apierrors.APIErrorResponse'
        "403":
          description: Forbidden error
          schema:
            $ref: '#/definitions/apierrors.APIErrorResponse'
        "404":
          description: Data not found error
          schema:
            $ref: '#/definitions/apierrors.APIErrorResponse'
        "409":
          description: Data conflict error
          schema:
            $ref: '#/definitions/apierrors.APIErrorResponse'
        "500":
          description: Internal server error
          schema:
            $ref: '#/definitions/apierrors.APIErrorResponse'
      summary: List Cadre Details
      tags:
      - Post Management
...
  /post-management/designations:
    get:
      consumes:
      - application/json
      description: Fetches the list of designations filtered by group and cadre codes.
        Also returns the complete list of designation types.
      parameters:
      - description: Group Code (optional)
        in: query
        name: group-code
        type: string
      - description: Cadre Code (optional)
        in: query
        name: cadre-code
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: Filtered designations and all designation types retrieved successfully
          schema:
            $ref: '#/definitions/response.ListAndFilterDesignationsAPIResponse'
        "400":
          description: Validation error
          schema:
            $ref: '#/definitions/apierrors.APIErrorResponse'
        "401":
          description: Unauthorized error
          schema:
            $ref: '#/definitions/apierrors.APIErrorResponse'
        "403":
          description: Forbidden error
          schema:
            $ref: '#/definitions/apierrors.APIErrorResponse'
        "404":
          description: Data not found error
          schema:
            $ref: '#/definitions/apierrors.APIErrorResponse'
        "409":
          description: Data conflict error
          schema:
            $ref: '#/definitions/apierrors.APIErrorResponse'
        "500":
          description: Internal server error
          schema:
            $ref: '#/definitions/apierrors.APIErrorResponse'
      summary: Get and Filter Designations
      tags:
      - Designation Management
...
  /post-management/files:
    get:
      consumes:
      - application/json
      description: Fetch document metadata for a given Office ID, download the files,
        and return them as a ZIP archive.
      parameters:
      - description: Office ID (unique identifier for the office)
        in: query
        name: office_id
        required: true
        type: integer
      produces:
      - application/zip
      responses:
        "200":
          description: OK
          schema:
            type: file
        "400":
          description: 'Bad Request: Invalid or missing Office ID'
          schema:
            $ref: '#/definitions/apierrors.APIErrorResponse'
        "404":
          description: 'Not Found: No files found for the specified Office ID'
          schema:
            $ref: '#/definitions/apierrors.APIErrorResponse'
        "500":
          description: 'Internal Server Error: Failed to fetch documents or generate
            ZIP'
          schema:
            $ref: '#/definitions/apierrors.APIErrorResponse'
      summary: Download Files as a ZIP for a specific Office ID
      tags:
      - Post Management
...
  /post-management/makers:
    get:
      consumes:
      - application/json
      description: Fetches posts for a specific Office ID and Maker. Can be filtered
        by Post ID.
      parameters:
      - description: Office ID
        in: query
        name: office-id
        required: true
        type: integer
      - description: Post ID (optional)
        in: query
        name: post-id
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: Posts retrieved successfully
          schema:
            $ref: '#/definitions/response.FetchPostsByOfficeIDAndMakerAPIResponse'
        "400":
          description: Validation error
          schema:
            $ref: '#/definitions/apierrors.APIErrorResponse'
        "401":
          description: Unauthorized error
          schema:
            $ref: '#/definitions/apierrors.APIErrorResponse'
        "403":
          description: Forbidden error
          schema:
            $ref: '#/definitions/apierrors.APIErrorResponse'
        "404":
          description: Data not found error
          schema:
            $ref: '#/definitions/apierrors.APIErrorResponse'
        "409":
          description: Data conflict error
          schema:
            $ref: '#/definitions/apierrors.APIErrorResponse'
        "500":
          description: Internal server error
          schema:
            $ref: '#/definitions/apierrors.APIErrorResponse'
      summary: Get Posts by Office ID and Maker
      tags:
      - Post Management
    post:
      consumes:
      - application/json
      description: Creates new post management maker records.
      parameters:
      - description: Request body for creating post management maker records
        in: body
        name: PostManagementMakerCreateRequests
        required: true
        schema:
          $ref: '#/definitions/handler.PostManagementMakerCreateRequests'
      produces:
      - application/json
      responses:
        "201":
          description: Post Management Maker Records Created Successfully
          schema:
            $ref: '#/definitions/response.CreatePostManagementMakerAPIResponse'
        "400":
          description: Validation error
          schema:
            $ref: '#/definitions/apierrors.APIErrorResponse'
        "401":
          description: Unauthorized error
          schema:
            $ref: '#/definitions/apierrors.APIErrorResponse'
        "403":
          description: Forbidden error
          schema:
            $ref: '#/definitions/apierrors.APIErrorResponse'
        "404":
          description: Data not found error
          schema:
            $ref: '#/definitions/apierrors.APIErrorResponse'
        "409":
          description: Data conflict error
          schema:
            $ref: '#/definitions/apierrors.APIErrorResponse'
        "500":
          description: Internal server error
          schema:
            $ref: '#/definitions/apierrors.APIErrorResponse'
      summary: Create Post Management Maker Records
      tags:
      - Post Management
...
  /post-management/office-establishment-register:
    get:
      consumes:
      - application/json
      description: Fetches establishment register details by either register ID or
        a combination of office ID and office type.
      parameters:
      - description: Register ID (optional)
        in: query
        name: register-id
        type: string
      - description: Office ID (required if register-id is not provided)
        in: query
        name: office-id
        type: string
      - description: Office Type (required if office-id is provided)
        in: query
        name: office-type
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: Establishment register details retrieved successfully
          schema:
            $ref: '#/definitions/response.FetchEstablishmentRegisterAPIResponse'
        "400":
          description: Validation error or missing parameters
          schema:
            $ref: '#/definitions/apierrors.APIErrorResponse'
        "401":
          description: Unauthorized error
          schema:
            $ref: '#/definitions/apierrors.APIErrorResponse'
        "403":
          description: Forbidden error
          schema:
            $ref: '#/definitions/apierrors.APIErrorResponse'
        "404":
          description: Data not found error
          schema:
            $ref: '#/definitions/apierrors.APIErrorResponse'
        "409":
          description: Data conflict error
          schema:
            $ref: '#/definitions/apierrors.APIErrorResponse'
        "500":
          description: Internal server error
          schema:
            $ref: '#/definitions/apierrors.APIErrorResponse'
      summary: Get Establishment Register
      tags:
      - Post Management
    post:
      consumes:
      - application/json
      description: Creates a new establishment register with the provided details.
      parameters:
      - description: Details for creating an establishment register
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/handler.EstablishmentRegisterRequest'
      produces:
      - application/json
      responses:
        "201":
          description: Establishment register created successfully
          schema:
            $ref: '#/definitions/response.CreateEstablishmentRegisterAPIResponse'
        "400":
          description: Validation error
          schema:
            $ref: '#/definitions/apierrors.APIErrorResponse'
        "401":
          description: Unauthorized error
          schema:
            $ref: '#/definitions/apierrors.APIErrorResponse'
        "403":
          description: Forbidden error
          schema:
            $ref: '#/definitions/apierrors.APIErrorResponse'
        "404":
          description: Data not found error
          schema:
            $ref: '#/definitions/apierrors.APIErrorResponse'
        "500":
          description: Internal server error
          schema:
            $ref: '#/definitions/apierrors.APIErrorResponse'
      summary: Create Establishment Register
      tags:
      - Post Management
...
  /post-management/posts:
    get:
      consumes:
      - application/json
      description: Retrieve post management details using various filters like Office
        ID, Post ID, Cadre Name, and Approved By.
      parameters:
      - description: Office ID (required)
        in: query
        name: office-id
        required: true
        type: integer
      - description: Post ID (optional)
        in: query
        name: post-id
        type: integer
      - description: Cadre Name (optional)
        in: query
        name: cadre-name
        type: string
      - description: Approved By (optional)
        in: query
        name: approved-by
        type: string
      - description: 'Number of records to skip (default: 0)'
        in: query
        name: skip
        type: integer
      - description: 'Number of records to retrieve (default: 10)'
        in: query
        name: limit
        type: integer
      - description: Field to sort by (e.g., id, name)
        in: query
        name: order_by
        type: string
      - description: Sort direction (asc or desc)
        in: query
        name: sort_type
        type: string
      - description: 'Flag to include total record count in the response (default:
          false)'
        in: query
        name: total_records_required
        type: boolean
      produces:
      - application/json
      responses:
        "200":
          description: Post management details retrieved successfully
          schema:
            $ref: '#/definitions/response.PostManagementByOfficeAndPostAPIResponse'
        "400":
          description: Validation error
          schema:
            $ref: '#/definitions/apierrors.APIErrorResponse'
        "401":
          description: Unauthorized error
          schema:
            $ref: '#/definitions/apierrors.APIErrorResponse'
        "403":
          description: Forbidden error
          schema:
            $ref: '#/definitions/apierrors.APIErrorResponse'
        "404":
          description: Data not found error
          schema:
            $ref: '#/definitions/apierrors.APIErrorResponse'
        "409":
          description: Data conflict error
          schema:
            $ref: '#/definitions/apierrors.APIErrorResponse'
        "500":
          description: Internal server error
          schema:
            $ref: '#/definitions/apierrors.APIErrorResponse'
      summary: Get Post Management by Office and Post
      tags:
      - Post Management
...
schemes:
- http
- https
swagger: "2.0"